MongoDB vs. CouchBase
The project is to compare the two distributed, document – oriented and NoSQL databases which has a share – nothing architecture and supports JSON. The installation of MongoDB is made by using MongoDB Atlas which is a cloud version. Both MongoDB and CouchBase were installed using EC2 in Amazon Web Services. 8GB of RAM was taken for each system. Two datasets were imported into both the systems that carries 192K documents in one dataset and 1.2M documents in another dataset. Queries that retrieves the same data written in MongoDB and CouchBase format and analyzed both systems performance and scalability. Joins, aggregate functions, range and point queries are taken to compare both the systems to show which is faster in performance and scalability.